Swapped some GPIO to avoid PCINT problems
This lead to a tasty basket weave of traces under the MCU, but that's
fine I guess.

The only reason this change was "necessary" is because regular arduino
doesn't expose gpio on the crystal pins. Therefore the PCINT libraries
don't seem to support enabling the pin change interrupt on those pins.

It was neat to have them as buttons, from a routing perspective, but
this should -work- which seems more important.